+++ package.json
@@ -6,11 +6,6 @@
     "@cloudscape-design/components": "^3.0.656",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.19",
-    "@mui/material": "^5.15.19",
-    "@mui/styled-engine-sc": "^6.0.0-alpha.18",
-    "@mui/styles": "^5.16.4",
-    "@mui/x-data-grid": "^7.6.1",
     "@reduxjs/toolkit": "^2.2.5",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -18,23 +13,18 @@
     "ajv": "^8.16.0",
     "ajv-keywords": "^5.1.0",
     "axios": "^1.7.4",
-    "chart.js": "^4.4.3",
     "html2canvas": "^1.4.1",
-    "html2pdf.js": "^0.10.2",
-    "immer": "^10.1.1",
     "jsbarcode": "^3.11.6",
     "jspdf": "^2.5.2",
     "jwt-decode": "^4.0.0",
     "react": "^18.3.1",
-    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.3.0",
+    "react-icons": "^5.5.0",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
-    "react-toastify": "^10.0.5",
+    "react-toastify": "^11.0.5",
     "sass-loader": "^14.2.1",
-    "simplebar-react": "^3.2.5",
     "styled-components": "^6.1.11",
     "web-vitals": "^2.1.4",
     "xlsx": "^0.18.5"
