--- src/Redux-Store/Products/dummy/orderHistory.json
@@ -1,163 +1,0 @@
-[
-    {
-      "order_id": "P001",
-      "status": "Unfulfilled",
-      "date": "Jun 15,2024",
-      "location": "PBEI",
-      "quantity": "500KG",
-      "subtotal": "Rs. 100"
-    },
-    {
-      "order_id": "P002",
-      "status": "Unpaid",
-      "date": "Jul 20,2024",
-      "location": "Kukatpally",
-      "quantity": "250KG",
-      "subtotal": "Rs. 50"
-    },
-    {
-      "order_id": "P003",
-      "status": "Uninvoiced",
-      "date": "Aug 25,2024",
-      "location": "Gachibowli",
-      "quantity": "750KG",
-      "subtotal": "Rs. 150"
-    },
-    {
-      "order_id": "P004",
-      "status": "Quote",
-      "date": "Sep 10,2024",
-      "location": "Vasati",
-      "quantity": "300KG",
-      "subtotal": "Rs. 60"
-    },
-    {
-      "order_id": "P005",
-      "status": "Fulfilled",
-      "date": "Oct 18,2024",
-      "location": "Shaikpet",
-      "quantity": "600KG",
-      "subtotal": "Rs. 120"
-    },
-    {
-      "order_id": "P006",
-      "status": "Draft",
-      "date": "Nov 22,2024",
-      "location": "Moosapet",
-      "quantity": "450KG",
-      "subtotal": "Rs. 90"
-    },
-    {
-      "order_id": "P007",
-      "status": "Stared",
-      "date": "Dec 15,2024",
-      "location": "PBEI",
-      "quantity": "800KG",
-      "subtotal": "Rs. 160"
-    },
-    {
-      "order_id": "P008",
-      "status": "Unfulfilled",
-      "date": "Jan 10,2024",
-      "location": "Vasati",
-      "quantity": "100KG",
-      "subtotal": "Rs. 20"
-    },
-    {
-      "order_id": "P009",
-      "status": "Unpaid",
-      "date": "Feb 05,2024",
-      "location": "Gachibowli",
-      "quantity": "350KG",
-      "subtotal": "Rs. 70"
-    },
-    {
-      "order_id": "P010",
-      "status": "Uninvoiced",
-      "date": "Mar 25,2024",
-      "location": "Moosapet",
-      "quantity": "500KG",
-      "subtotal": "Rs. 100"
-    },
-    {
-      "order_id": "P011",
-      "status": "Quote",
-      "date": "Apr 30,2024",
-      "location": "Shaikpet",
-      "quantity": "750KG",
-      "subtotal": "Rs. 150"
-    },
-    {
-      "order_id": "P012",
-      "status": "Fulfilled",
-      "date": "May 12,2024",
-      "location": "Kukatpally",
-      "quantity": "900KG",
-      "subtotal": "Rs. 180"
-    },
-    {
-      "order_id": "P013",
-      "status": "Draft",
-      "date": "Jun 16,2024",
-      "location": "PBEI",
-      "quantity": "200KG",
-      "subtotal": "Rs. 40"
-    },
-    {
-      "order_id": "P014",
-      "status": "Stared",
-      "date": "Jul 18,2024",
-      "location": "Vasati",
-      "quantity": "650KG",
-      "subtotal": "Rs. 130"
-    },
-    {
-      "order_id": "P015",
-      "status": "Unfulfilled",
-      "date": "Aug 20,2024",
-      "location": "Shaikpet",
-      "quantity": "550KG",
-      "subtotal": "Rs. 110"
-    },
-    {
-      "order_id": "P016",
-      "status": "Unpaid",
-      "date": "Sep 22,2024",
-      "location": "Gachibowli",
-      "quantity": "700KG",
-      "subtotal": "Rs. 140"
-    },
-    {
-      "order_id": "P017",
-      "status": "Uninvoiced",
-      "date": "Oct 24,2024",
-      "location": "Moosapet",
-      "quantity": "300KG",
-      "subtotal": "Rs. 60"
-    },
-    {
-      "order_id": "P018",
-      "status": "Quote",
-      "date": "Nov 26,2024",
-      "location": "PBEI",
-      "quantity": "450KG",
-      "subtotal": "Rs. 90"
-    },
-    {
-      "order_id": "P019",
-      "status": "Fulfilled",
-      "date": "Dec 28,2024",
-      "location": "Kukatpally",
-      "quantity": "800KG",
-      "subtotal": "Rs. 160"
-    },
-    {
-      "order_id": "P020",
-      "status": "Draft",
-      "date": "Jan 01,2024",
-      "location": "Shaikpet",
-      "quantity": "150KG",
-      "subtotal": "Rs. 30"
-    }
-  ]
-
